third_party/WebKit/LayoutTests/media/track/vtt-cue-exceptions.html

<!DOCTYPE html>
<title>Tests that VTTCue exceptions are properly messaged to developers.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function getExceptionMsg(property, value) {
    return "Failed to set the '" + property
        + "' property on 'VTTCue': The value provided ("
        + value + ") is outside the range [0, 100].";
}
var cue = new VTTCue(0, 0, "Test.");
function testProperty(property, value) {
    var expected_exceptioin_msg = getExceptionMsg(property, value);
    test(function() {
        assert_throws("IndexSizeError", function() { cue[property] = value; });
        try {
            cue[property] = value;
        } catch (e) {
            assert_equals(e.message, expected_exceptioin_msg);
        }
    }, expected_exceptioin_msg);
}
testProperty("position", -1);
testProperty("position", 101);
testProperty("size", -1);
testProperty("size", 101);
</script>
